+++ src/test/groovy/ru/vyarus/dropwizard/orient/StartApplicationTest.groovy
@@ -35,5 +35,15 @@ class StartApplicationTest extends AbstractTest {
         data = new URL("http://localhost:8081/healthcheck").getText()
         then: "all good"
         data && data.contains('"orient-server":{"healthy":true,"message":"OK"}')
+
+        when: "accessing orient servlet"
+        data = new URL("http://localhost:8081/orient/").getText()
+        then: "all good"
+        data != null
+
+        when: "accessing studio through orient servlet"
+        data = new URL("http://localhost:8081/orient/studio").getText()
+        then: "all good"
+        data != null
     }
 }